# apps/api/openapi.json
{"openapi":"3.0.0","paths":{"/orders":{"post":{"operationId":"OrderController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderDto"}}}},"responses":{"201":{"description":""}},"tags":["orders"]},"get":{"operationId":"OrderController_findMine","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["orders"]}},"/orders/track":{"get":{"operationId":"OrderController_track","parameters":[{"name":"orderNumber","required":true,"in":"query","schema":{"type":"string"}},{"name":"email","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["orders"]}},"/orders/{id}":{"get":{"operationId":"OrderController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["orders"]}},"/admin/orders":{"get":{"operationId":"AdminOrderController_findAll","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-orders"]}},"/admin/orders/{id}/status":{"patch":{"operationId":"AdminOrderController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-orders"]}},"/checkout/preview":{"post":{"operationId":"PricingController_preview","parameters":[],"responses":{"201":{"description":""}},"tags":["checkout"]}},"/shipping/zones/{countryCode}/methods":{"get":{"operationId":"ShippingController_getMethods","parameters":[{"name":"countryCode","required":true,"in":"path","schema":{"type":"string"}},{"name":"subtotal","required":true,"in":"query","schema":{"type":"string"}},{"name":"city","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["shipping"]}},"/currencies":{"get":{"operationId":"CurrencyController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["currency"]}},"/admin/inventory/products/{id}":{"get":{"operationId":"InventoryController_getProductInventory","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["inventory"]}},"/admin/inventory/low-stock":{"get":{"operationId":"InventoryController_getLowStock","parameters":[{"name":"threshold","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["inventory"]}},"/admin/inventory/adjustments":{"post":{"operationId":"InventoryController_adjust","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["inventory"]}},"/notifications/devices":{"post":{"operationId":"NotificationController_registerDevice","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["notifications"]}},"/admin/reports/overview":{"get":{"operationId":"ReportingController_overview","parameters":[{"name":"from","required":true,"in":"query","schema":{"type":"string"}},{"name":"to","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-reports"]}},"/admin/reports/sales":{"get":{"operationId":"ReportingController_sales","parameters":[{"name":"groupBy","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-reports"]}},"/users":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/users/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/me":{"get":{"operationId":"UserMeController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["UserMe"]},"patch":{"operationId":"UserMeController_updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":""}},"tags":["UserMe"]},"delete":{"operationId":"UserMeController_deleteMe","parameters":[],"responses":{"200":{"description":""}},"tags":["UserMe"]}},"/auth/sign-up":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/sign-in":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/logout":{"post":{"operationId":"AuthController_logout","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/refresh":{"post":{"operationId":"AuthController_refresh","parameters":[],"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["auth"]}},"/auth/reset-password":{"post":{"operationId":"AuthController_resetPassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/verify-code":{"post":{"operationId":"AuthController_verifyCode","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyEmailAndCode"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/auth/change-password":{"post":{"operationId":"AuthController_changePassword","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SigninDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"]}},"/category":{"get":{"operationId":"CategoryController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Category"]},"post":{"operationId":"CategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["Category"]}},"/category/{id}":{"get":{"operationId":"CategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]},"patch":{"operationId":"CategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["Category"]},"delete":{"operationId":"CategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Category"]}},"/sub-category":{"get":{"operationId":"SubCategoryController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["SubCategory"]},"post":{"operationId":"SubCategoryController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSubCategoryDto"}}}},"responses":{"201":{"description":""}},"tags":["SubCategory"]}},"/sub-category/{id}":{"get":{"operationId":"SubCategoryController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["SubCategory"]},"patch":{"operationId":"SubCategoryController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSubCategoryDto"}}}},"responses":{"200":{"description":""}},"tags":["SubCategory"]},"delete":{"operationId":"SubCategoryController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["SubCategory"]}},"/brand":{"get":{"operationId":"BrandController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Brand"]},"post":{"operationId":"BrandController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBrandDto"}}}},"responses":{"201":{"description":""}},"tags":["Brand"]}},"/brand/{id}":{"get":{"operationId":"BrandController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Brand"]},"patch":{"operationId":"BrandController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBrandDto"}}}},"responses":{"200":{"description":""}},"tags":["Brand"]},"delete":{"operationId":"BrandController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Brand"]}},"/coupon":{"get":{"operationId":"CouponController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Coupon"]},"post":{"operationId":"CouponController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCouponDto"}}}},"responses":{"201":{"description":""}},"tags":["Coupon"]}},"/coupon/{id}":{"get":{"operationId":"CouponController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Coupon"]},"patch":{"operationId":"CouponController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCouponDto"}}}},"responses":{"200":{"description":""}},"tags":["Coupon"]},"delete":{"operationId":"CouponController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Coupon"]}},"/supplier":{"post":{"operationId":"SupplierController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSupplierDto"}}}},"responses":{"201":{"description":""}},"tags":["Supplier"]},"get":{"operationId":"SupplierController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Supplier"]}},"/supplier/{id}":{"get":{"operationId":"SupplierController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Supplier"]},"patch":{"operationId":"SupplierController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSupplierDto"}}}},"responses":{"200":{"description":""}},"tags":["Supplier"]},"delete":{"operationId":"SupplierController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Supplier"]}},"/request-product":{"get":{"operationId":"RequestProductController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["RequestProduct"]},"post":{"operationId":"RequestProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRequestProductDto"}}}},"responses":{"201":{"description":""}},"tags":["RequestProduct"]}},"/request-product/{id}":{"get":{"operationId":"RequestProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["RequestProduct"]},"patch":{"operationId":"RequestProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRequestProductDto"}}}},"responses":{"200":{"description":""}},"tags":["RequestProduct"]},"delete":{"operationId":"RequestProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["RequestProduct"]}},"/tax":{"get":{"operationId":"TaxController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Tax"]},"post":{"operationId":"TaxController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTaxDto"}}}},"responses":{"201":{"description":""}},"tags":["Tax"]}},"/tax/{id}":{"get":{"operationId":"TaxController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tax"]},"patch":{"operationId":"TaxController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTaxDto"}}}},"responses":{"200":{"description":""}},"tags":["Tax"]},"delete":{"operationId":"TaxController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Tax"]}},"/product":{"get":{"operationId":"ProductController_findAll","parameters":[],"responses":{"200":{"description":""}},"tags":["Product"]},"post":{"operationId":"ProductController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProductDto"}}}},"responses":{"201":{"description":""}},"tags":["Product"]}},"/product/{id}":{"get":{"operationId":"ProductController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"]},"patch":{"operationId":"ProductController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProductDto"}}}},"responses":{"200":{"description":""}},"tags":["Product"]},"delete":{"operationId":"ProductController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Product"]}},"/review/product/{id}":{"get":{"operationId":"ReviewController_findAll","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]}},"/review/user/{id}":{"get":{"operationId":"ReviewController_findByUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]}},"/review":{"post":{"operationId":"ReviewController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReviewDto"}}}},"responses":{"201":{"description":""}},"tags":["Review"]}},"/review/{id}":{"patch":{"operationId":"ReviewController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateReviewDto"}}}},"responses":{"200":{"description":""}},"tags":["Review"]},"delete":{"operationId":"ReviewController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Review"]}},"/cart/guest":{"post":{"operationId":"CartController_createGuestCart","parameters":[],"responses":{"201":{"description":""}},"tags":["cart"]}},"/cart":{"get":{"operationId":"CartController_getCart","parameters":[],"responses":{"200":{"description":""}},"tags":["cart"]}},"/cart/items":{"post":{"operationId":"CartController_addItem","parameters":[],"responses":{"201":{"description":""}},"tags":["cart"]}},"/cart/merge":{"post":{"operationId":"CartController_merge","parameters":[],"responses":{"201":{"description":""}},"tags":["cart"]}},"/catalog/products":{"get":{"operationId":"CatalogController_findAll","parameters":[{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"limit","required":true,"in":"query","schema":{"type":"string"}},{"name":"keyword","required":true,"in":"query","schema":{"type":"string"}},{"name":"accept-language","required":true,"in":"header","schema":{"type":"string"}},{"name":"x-currency","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["catalog"]}},"/catalog/products/{slug}":{"get":{"operationId":"CatalogController_findBySlug","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}},{"name":"accept-language","required":true,"in":"header","schema":{"type":"string"}},{"name":"x-currency","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["catalog"]}},"/payments/kashier/session":{"post":{"operationId":"PaymentController_createSession","parameters":[],"responses":{"201":{"description":""}},"tags":["payments"]}},"/payments/kashier/webhook":{"post":{"operationId":"PaymentController_webhook","parameters":[{"name":"x-kashier-signature","required":true,"in":"header","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"tags":["payments"]}},"/payments/kashier/callback":{"get":{"operationId":"PaymentController_callback","parameters":[{"name":"orderId","required":true,"in":"query","schema":{"type":"string"}},{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["payments"]}},"/payments/cod/confirm":{"post":{"operationId":"PaymentController_confirmCod","parameters":[],"responses":{"201":{"description":""}},"tags":["payments"]}},"/returns":{"post":{"operationId":"ReturnsController_create","parameters":[],"responses":{"201":{"description":""}},"tags":["returns"]}},"/returns/order/{orderId}":{"get":{"operationId":"ReturnsController_findByOrder","parameters":[{"name":"orderId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["returns"]}},"/admin/returns":{"get":{"operationId":"AdminReturnsController_findAll","parameters":[{"name":"status","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-returns"]}},"/admin/returns/{id}/status":{"patch":{"operationId":"AdminReturnsController_updateStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-returns"]}},"/admin/campaigns":{"post":{"operationId":"CampaignController_create","parameters":[],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]},"get":{"operationId":"CampaignController_findAll","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]}},"/admin/campaigns/{id}":{"get":{"operationId":"CampaignController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]},"patch":{"operationId":"CampaignController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]},"delete":{"operationId":"CampaignController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]}},"/admin/campaigns/{id}/send":{"post":{"operationId":"CampaignController_send","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]}},"/admin/campaigns/{id}/stats":{"get":{"operationId":"CampaignController_stats","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"tags":["admin-campaigns"]}},"/health":{"get":{"operationId":"HealthController_check","parameters":[],"responses":{"200":{"description":""}},"tags":["health"]}}},"info":{"title":"EE Commerce API","description":"Production e-commerce API","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"OrderItemDto":{"type":"object","properties":{"productId":{"type":"string"},"variantSku":{"type":"string"},"title":{"type":"string"},"quantity":{"type":"number"},"unitPrice":{"type":"number"}},"required":["productId","title","quantity","unitPrice"]},"AddressDto":{"type":"object","properties":{"fullName":{"type":"string"},"phone":{"type":"string"},"country":{"type":"string"},"city":{"type":"string"},"street":{"type":"string"},"postalCode":{"type":"string"}},"required":["fullName","phone","country","city","street"]},"CreateOrderDto":{"type":"object","properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDto"}},"shippingAddress":{"$ref":"#/components/schemas/AddressDto"},"billingAddress":{"$ref":"#/components/schemas/AddressDto"},"paymentMethod":{"type":"string","enum":["cod","kashier"]},"guestEmail":{"type":"string"},"guestPhone":{"type":"string"},"shippingMethodId":{"type":"string"},"couponCode":{"type":"string"},"currency":{"type":"string"},"notes":{"type":"string"}},"required":["items","shippingAddress","paymentMethod"]},"CreateUserDto":{"type":"object","properties":{}},"UpdateUserDto":{"type":"object","properties":{}},"SignupDto":{"type":"object","properties":{}},"SigninDto":{"type":"object","properties":{}},"VerifyEmailDto":{"type":"object","properties":{}},"VerifyEmailAndCode":{"type":"object","properties":{}},"CreateCategoryDto":{"type":"object","properties":{}},"UpdateCategoryDto":{"type":"object","properties":{}},"CreateSubCategoryDto":{"type":"object","properties":{}},"UpdateSubCategoryDto":{"type":"object","properties":{}},"CreateBrandDto":{"type":"object","properties":{}},"UpdateBrandDto":{"type":"object","properties":{}},"CreateCouponDto":{"type":"object","properties":{}},"UpdateCouponDto":{"type":"object","properties":{}},"CreateSupplierDto":{"type":"object","properties":{}},"UpdateSupplierDto":{"type":"object","properties":{}},"CreateRequestProductDto":{"type":"object","properties":{}},"UpdateRequestProductDto":{"type":"object","properties":{}},"CreateTaxDto":{"type":"object","properties":{}},"UpdateTaxDto":{"type":"object","properties":{}},"CreateProductDto":{"type":"object","properties":{}},"UpdateProductDto":{"type":"object","properties":{}},"CreateReviewDto":{"type":"object","properties":{}},"UpdateReviewDto":{"type":"object","properties":{}}}}}